Refactor read_scaffold_graph to ntlink_utils.py
